pyproject.toml
[project]
name = "fastapi-oauth-client"
version = "0.1.0"
readme = "README.md"
authors = [{ name = "Ashwin A Nayar", email = "ashwin5059198@gmail.com" }]
requires-python = ">=3.10"
dependencies = ["fastapi[standard-no-fastapi-cloud-cli]", "authlib", "pyjwt"]

[build-system]
requires = ["uv_build>=0.10.2,<0.11.0"]
build-backend = "uv_build"

[tool.ruff]
target-version = "py310"
line-length = 120

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "D10",
    "ANN002",
    "ANN003",
    "D203",   # incompatible with D211
    "D213",   # incompatible with D212
    "COM812", # redundant when used along with `ruff format`
    "INP",
]
